requirements.txt
pyaudio==0.2.13
numpy==1.24.3
pyserial==3.5 